pyproject.toml
[project]
name = "grow_api_python_example"
version = "0.0.1"
authors = [{ name = "bjorn", email = "bjorn@photosynthetic.com" }]
description = "Grow api example in python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    'psutil',
    'appdirs',
    'numpy',
    'protobuf',
    'ordered-set',
    'betterproto[compiler]==2.0.0b7',
    'betterproto==2.0.0b7',
    'grpcio-tools',
    'asyncio',
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/grow_api_python_example"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

